Fix invalid html code in about
Auto closing spans are not a valid syntax.
By the way, I didn't really get why this was needed. To link directly to this part of the page? Did not see any link to do that.

I move the id to the following div at the moment, but I think this can be remove every there is no link to it.